refactor: Round 14 - extract CodebaseHealth modules (885→89 LOC, -90%)
Extract focused modules from CodebaseHealth.tsx:
- types.ts: Shared types and constants
- StatusSelector.tsx: Status dropdown component
- StatsTab.tsx: Stats dashboard tab
- FilesTab.tsx: File inventory tab
- TodosTab.tsx: Refactor todo list tab
- Legend.tsx: Size/status/effort legend
- index.ts: Module re-exports

Main file reduced from 885 to 89 LOC.
