+++ web/src/pages/CodebaseHealth/index.ts
@@ -1,0 +1,9 @@
+/**
+ * CodebaseHealth module exports
+ */
+export { StatusSelector } from './StatusSelector';
+export { StatsTab } from './StatsTab';
+export { FilesTab } from './FilesTab';
+export { TodosTab } from './TodosTab';
+export { Legend } from './Legend';
+export * from './types';
